# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Forum Group Discussion",
        "title": "Forum Group Discussion API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kelompok 22",
            "url": "http://www.swagger.io/support",
            "email": "erikriosetiawan15@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "moot-rest-api.herokuapp.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/categories": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "This endpoint is used to get all category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get Categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.categoriesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint is used to create a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a Category",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "default",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.CreateCategory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.createThreadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKey": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint is used to update a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update a Category",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "default",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UpdateCategory"
                        }
                    },
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKey": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint is used to delete a category by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Delete Category by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/categories/{id}/threads": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "This endpoint is used to get the threads of particular category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get Category Threads",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page, default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit, default 10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.threadsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.categoriesData": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Category"
                    },
                    "x-order": "0"
                }
            }
        },
        "controller.categoriesResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "x-order": "0"
                },
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "x-order": "2",
                    "$ref": "#/definitions/controller.categoriesData"
                }
            }
        },
...
        "controller.createThreadResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "x-order": "0"
                },
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "x-order": "2",
                    "$ref": "#/definitions/controller.idData"
                }
            }
        },
        "controller.idData": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                }
            }
        },
...
        "controller.pageInfoData": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "x-order": "0"
                },
                "page": {
                    "type": "integer",
                    "x-order": "1"
                },
                "pageTotal": {
                    "type": "integer",
                    "x-order": "2"
                },
                "total": {
                    "type": "integer",
                    "x-order": "3"
                }
            }
        },
...
        "controller.threadsData": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string",
                    "x-order": "0"
                },
                "title": {
                    "type": "string",
                    "x-order": "1"
                },
                "categoryID": {
                    "type": "string",
                    "x-order": "10"
                },
                "categoryName": {
                    "type": "string",
                    "x-order": "11"
                },
                "publishedOn": {
                    "description": "PublishedOn layout format: time.RFC822 (02 Jan 06 15:04 MST)",
                    "type": "string",
                    "x-order": "12"
                },
                "isLiked": {
                    "type": "boolean",
                    "x-order": "13"
                },
                "isFollowed": {
                    "type": "boolean",
                    "x-order": "14"
                },
                "description": {
                    "type": "string",
                    "x-order": "2"
                },
                "totalViewer": {
                    "type": "integer",
                    "x-order": "3"
                },
                "totalLike": {
                    "type": "integer",
                    "x-order": "4"
                },
                "totalFollower": {
                    "type": "integer",
                    "x-order": "5"
                },
                "totalComment": {
                    "type": "integer",
                    "x-order": "6"
                },
                "creatorID": {
                    "type": "string",
                    "x-order": "7"
                },
                "creatorUsername": {
                    "type": "string",
                    "x-order": "8"
                },
                "creatorName": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
        "controller.threadsInfoWrapper": {
            "type": "object",
            "properties": {
                "threads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.threadsData"
                    },
                    "x-order": "0"
                },
                "pageInfo": {
                    "x-order": "1",
                    "$ref": "#/definitions/controller.pageInfoData"
                }
            }
        },
        "controller.threadsResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "x-order": "0"
                },
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "x-order": "2",
                    "$ref": "#/definitions/controller.threadsInfoWrapper"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "payload.CreateCategory": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "x-order": "0"
                },
                "description": {
                    "type": "string",
                    "minLength": 2,
                    "x-order": "1"
                }
            }
        },
...
        "payload.UpdateCategory": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "x-order": "0"
                },
                "description": {
                    "type": "string",
                    "minLength": 2,
                    "x-order": "1"
                }
            }
        },
...
        "response.Category": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string",
                    "x-order": "0"
                },
                "name": {
                    "type": "string",
                    "x-order": "1"
                },
                "description": {
                    "type": "string",
                    "x-order": "2"
                },
                "createdOn": {
                    "description": "CreatedOn layout format: time.RFC822 (02 Jan 06 15:04 MST)",
                    "type": "string",
                    "x-order": "3"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "API-Key",
            "in": "header"
        },
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}